+++ watchme/exporters/__init__.py
@@ -1,11 +1,11 @@
 '''
 
-Exporter class : defines properties of an exporter object
-Note : No validation is currently done for the exporters.
+Copyright (C) 2019 Vanessa Sochat
+Copyright (C) 2019 Antoine Solnichkin
 
-Author : Antoine Solnichkin
-Date : 04/05/2019
-
+This Source Code Form is subject to the terms of the
+Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
+with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 '''
 
@@ -54,4 +54,3 @@ class ExporterBase(object):
             if param not in self.params:
                 bot.error('Missing required parameter: %s' % param)
                 self.valid = False
-